Fixes for feedback in pull request; coverage still too low.
